requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==2.1.4
pyyaml==6.0.1
python-dateutil==2.8.2
schedule==1.2.0
